v1.4.0

+++ src/Service/Storage/DocumentStorageService.php
@@ -45,8 +45,10 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
  * ThumbnailStorage service. This allows us to separate the storage of the document and the storage of the thumbnails if needed.
  *
  * @SuppressWarnings(TooManyPublicMethods)
+ * @SuppressWarnings(CouplingBetweenObjects)
+ * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  */
-class DocumentStorageService
+class DocumentStorageService implements StorageAliveInterface
 {
     protected FilesystemOperator $storage;
     protected EntityManagerInterface $doctrine;
@@ -235,7 +237,7 @@ class DocumentStorageService
         return $this->retrieveResource($remotePath);
     }
 
-    public function storeDocument(\SplFileInfo $localFile, EntityWithFileInfo $document): bool
+    public function storeDocument(\SplFileInfo $localFile, EntityWithFileInfo $document, bool $flush = true): bool
     {
         $remotePath = $this->generateDocumentPath($document, $localFile);
 
@@ -254,7 +256,10 @@ class DocumentStorageService
         $file->setUploaded(true);
 
         $this->doctrine->persist($document);
-        $this->doctrine->flush();
+
+        if ($flush) {
+            $this->doctrine->flush();
+        }
 
         return true;
     }
@@ -405,6 +410,11 @@ class DocumentStorageService
     public function deleteAllFilesForDocument(Document $document): bool
     {
         try {
+            // Skip if file is not uploaded
+            if (! $document->getFileInfo()->isUploaded()) {
+                return true;
+            }
+
             $path = $this->generateDocumentPath($document, new \SplFileInfo($document->getFileInfo()->getPath() ?? ''));
             $this->storage->delete($path);
 
@@ -414,6 +424,38 @@ class DocumentStorageService
             }
         } catch (\Throwable $e) {
             $this->logger->error('Could not delete files from storage', [
+                'exception' => $e->getMessage(),
+                'path' => $path ?? '',
+            ]);
+
+            return false;
+        }
+
+        return true;
+    }
+
+    public function isAlive(): bool
+    {
+        $suffix = hash('sha256', random_bytes(32));
+
+        try {
+            $this->storage->write("healthcheck.{$suffix}", $suffix);
+            $content = $this->storage->read("healthcheck.{$suffix}");
+            $this->storage->delete("healthcheck.{$suffix}");
+        } catch (\Exception) {
+            return false;
+        }
+
+        return $content == $suffix;
+    }
+
+    public function removeFileForEntity(EntityWithFileInfo $entity): bool
+    {
+        try {
+            $path = $this->generateDocumentPath($entity, new \SplFileInfo($entity->getFileInfo()->getPath() ?? ''));
+            $this->storage->delete($path);
+        } catch (\Throwable $e) {
+            $this->logger->error('Could not delete file from storage for entity', [
                 'exception' => $e->getMessage(),
                 'path' => $path ?? '',
             ]);
